Fix C23 build, error "unterminated-string-initialization"
In C23, a string literal must fit in its array including its NUL terminator.

In version.c, the comment says there should not be a NUL terminator in SBAT data. But it seems there was one, at least in sbat.h. So either this is an old bug, or the comment should be changed.

Fix typo in xenver.h.